requirements.txt
python-dotenv
pandas
boto3
requests
python-docx

